// frontend/src/features/projects/types.ts
export enum ProjectMemberRole {
  Owner = 1,
  Member = 2,
  Viewer = 3,
}
// ...
export interface ProjectMemberDto {
  userId: string;
  displayName: string;
  email: string;
  addedAt: string;
  role?: ProjectMemberRole;
}